+++ R/spp_distribution.R
@@ -3,17 +3,17 @@
 #' Retrieve distributions data available for a given taxon concept for which the
 #' the taxon identifier is known.
 #'
-#' @param tax_id character string containing a species' taxon concept identifier
-#' (see \code{\link[rcites]{sppplus_taxonconcept}}).
+#' @param taxon_id character string containing a species' taxon concept identifier
+#' (see \code{\link[rcites]{spp_taxonconcept}}).
 #' @param collapse_tags a string used to collapse tags. Default is set to \code{NULL} meaning that tags column's elements remains lists.
 #' @param simplify a logical. Should the output be simplified? In other words,
 #' should columns of data.table objects returned be unlisted when they are
 #' lists made of single elements?
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
-#' \code{NULL} and requires the environment variable \code{SPPPLUS_TOKEN} to be
-#' set directly in \code{Renviron}. Alternatively \code{sppplus_login()} can
-#' be used to set \code{SPPPLUS_TOKEN} for the current session.
+#' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
+#' set directly in \code{Renviron}. Alternatively \code{set_token()} can
+#' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
 #'
 #' @return A data table with all distribution information.
 #'
@@ -25,17 +25,17 @@
 #'
 #' @examples
 #' \donttest{
-#'  res1 <- taxon_distribution(tax_id = '4521')
-#'  res2 <- taxon_distribution(tax_id = '4521', collapse_tags = ' + ', simplify = T)
+#'  res1 <- spp_distribution(taxon_id = '4521')
+#'  res2 <- spp_distribution(taxon_id = '4521', collapse_tags = ' + ', simplify = T)
 #' }
 
-taxon_distribution <- function(tax_id, collapse_tags = NULL, simplify = FALSE, token = NULL) {
+spp_distribution <- function(taxon_id, collapse_tags = NULL, simplify = FALSE, token = NULL) {
     # token check
     if (is.null(token)) 
-        token <- sppplus_getsecret()
+        token <- rcites_getsecret()
     # 
-    q_url <- sppplus_url(paste0("taxon_concepts/", tax_id, "/distributions.json"))
-    res <- sppplus_res(q_url, token)
+    q_url <- rcites_url(paste0("taxon_concepts/", taxon_id, "/distributions.json"))
+    res <- rcites_res(q_url, token)
     # get a data.table; tags and references are lists.
     out <- as.data.table(do.call(rbind, lapply(res, rbind)))
     if (!is.null(collapse_tags)) 
@@ -43,7 +43,7 @@ taxon_distribution <- function(tax_id, collapse_tags = NULL, simplify = FALSE, t
             paste(unlist(x), collapse = collapse_tags))
     ## 
     if (isTRUE(simplify)) 
-        sppplus_simplify(out)
+      rcites_simplify(out)
     # output
     out
 }