+++ R/spp_taxonconcept.R
@@ -8,16 +8,16 @@
 #' CITES appendix information are the only data returned (default is \code{FALSE}).
 #' @param token a character string containing the authentification token, see
 #' \url{https://api.speciesplus.net/documentation}. Default is set to
-#' \code{NULL} and requires the environment variable \code{SPPPLUS_TOKEN} to be
-#' set directly in \code{Renviron}. Alternatively, \code{sppplus_login()} can
-#' be used to set \code{SPPPLUS_TOKEN} for the current session.
+#' \code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
+#' set directly in \code{Renviron}. Alternatively, \code{set_token()} can
+#' be used to set \code{SPECIESPLUS_TOKEN} for the current session.
 #'
 #' @return
 #' If \code{appendix_only} is \code{TRUE}, then a data frame with a species'
 #' taxon id and CITES appendix information is returned. Otherwise, a list of
 #' objects of class \code{data.table} with all CITES Species+ taxon_concept
 #' information is returned. Importantly enough, this functions returns the
-#' taxon concept identifier that is required by the \code{taxon_*()} functions.
+#' taxon concept identifier that is required by the \code{spp_*()} functions.
 #'
 #' @references
 #' \url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
@@ -27,19 +27,19 @@
 #'
 #' @examples
 #' \donttest{
-#' res1 <- sppplus_taxonconcept(query_taxon = 'Loxodonta africana')
-#' res2 <- sppplus_taxonconcept(query_taxon = 'Loxodonta africana', appendix_only = TRUE)
-#' res3 <- sppplus_taxonconcept(query_taxon = 'Amazilia versicolor')
+#' res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
+#' res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', appendix_only = TRUE)
+#' res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
 #' }
 
-sppplus_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL) {
+spp_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NULL) {
     # token check
     if (is.null(token))
-        token <- sppplus_getsecret()
+        token <- rcites_getsecret()
     query <- gsub(pattern = " ", replacement = "%20", x = query_taxon)
     #
-    q_url <- sppplus_url(paste0("taxon_concepts.json", "?name=", query))
-    res <- sppplus_res(q_url, token)
+    q_url <- rcites_url(paste0("taxon_concepts.json", "?name=", query))
+    res <- rcites_res(q_url, token)
 
     if (!res$pagination$total_entries) {
         warning("Taxon not listed.")
@@ -49,7 +49,7 @@ sppplus_taxonconcept <- function(query_taxon, appendix_only = FALSE, token = NUL
             "active", "cites_listing")
         tmp <- as.data.table(do.call(rbind, lapply(res$taxon_concepts, function(x) rbind(x[nm]))))
         #
-        sppplus_simplify(tmp)
+        rcites_simplify(tmp)
         #
         if (isTRUE(appendix_only)) {
             out <- tmp