+++ tests/testthat/test-legislations.R
@@ -4,11 +4,11 @@ skip_on_cran()
 skip_if_no_auth()
 
 ut_pause()
-res1 <- taxon_eu_legislation(tax_id = tx_id)
+res1 <- spp_eu_legislation(taxon_id = tx_id)
 ut_pause()
-res2 <- taxon_eu_legislation(tax_id = tx_id, type ='listings')
+res2 <- spp_eu_legislation(taxon_id = tx_id, type ='listings')
 ut_pause()
-res3 <- taxon_eu_legislation(tax_id = tx_id, simplify = TRUE)
+res3 <- spp_eu_legislation(taxon_id = tx_id, simplify = TRUE)
 #
 nm1 <- c("id", "taxon_concept_id", "is_current", "annex", "change_type", "effective_at", "annotation")
 nm2 <- c("id", "taxon_concept_id", "notes", "start_date", "is_current", "eu_decision_type", "geo_entity", "start_event", "source", "term")
@@ -45,11 +45,11 @@ test_that("expected values", {
 context("legislation cites")
 
 ut_pause()
-res4 <- taxon_cites_legislation(tax_id = tx_id)
+res4 <- spp_cites_legislation(taxon_id = tx_id)
 ut_pause()
-res5 <- taxon_cites_legislation(tax_id = tx_id, type ='listings')
+res5 <- spp_cites_legislation(taxon_id = tx_id, type ='listings')
 ut_pause()
-res6 <- taxon_cites_legislation(tax_id = tx_id, simplify = TRUE)
+res6 <- spp_cites_legislation(taxon_id = tx_id, simplify = TRUE)
 
 test_that("expected output classes", {
   expect_equal(class(res4), "list")